Remove non-ASCII characters from nicknames
Details are included in the comment I wrote. I'm not sure if this is only an
issue on Linux, but I'm fixing this anyway since... well, I play this game on
Linux.

Definitely not a locale issue on my end. Most likely related to how the
results.csv file is encoded. Oh well.