Clean up nicknames more
Someone in NA had a CR (Carriage Return, '\r') character at the end of their
nickname, and that messed up the script when it read the results.csv file. The
lines that contained their nickname would essentially be cut in the middle, and
throw a "Line N has been tampered with" error.

The script was already removing newline characters from each line, but only at
the end. Now it removes newline characters in any position. Also, now it only
counts LF (Line Feed, '\n') as being the end of the line, not CR. This *can*
break in the future if someone ends up putting a '\n' character in their name
somehow. Maybe I should use the `csv` library to parse the results file...